backend/requirements.txt
# Required Python Version (3.12) ! #

# Web Framework & Web Server
Flask==3.0.2
gunicorn==23.0.0
Flask-Cors==4.0.0
Flask-SocketIO=5.5.1
Werkzeug==3.1.3

# Flask Database (PostgreSQL)
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.31
psycopg2-binary==2.9.10
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.7.1

# Deep Learning Libraries
torch==2.6.0
torchvision==0.21.0

# YOLO
ultralytics==8.3.0

# Image Processing Libraries
Pillow==11.1.0
opencv-python-headless==4.9.0.80
opencv-python==4.11.0.86

# Data handling and processing
numpy==1.26.4
pandas==2.2.1

# Machine Learning and Evaluation Metrics
scikit-learn==1.4.1.post1
scipy==1.12.0

# Visualization Tools
matplotlib==3.8.3

# Environment Variables Management
python-dotenv==1.0.1

# Logging & Monitoring
loguru==0.7.2

# Testing Libraries
pytest==8.1.1
pytest-flask==1.3.0
